pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hat"
version = "1.0.0"
description = "Company context switcher"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "pyyaml>=6.0",
    "keyring>=25.0",
    "httpx>=0.27",
]

[project.scripts]
hat = "hat.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/hat"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
